fix: move image-upload to components directory
